put visualizationcontrollertest and  required annatation on service and repository
