+++ src/main/java/com/revature/app/dao/VisualizationDao.java
@@ -1,5 +1,12 @@
 package com.revature.app.dao;
 
-public class VisualizationDao {
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+import com.revature.app.model.Visualization;
+
+@Repository
+public interface VisualizationDao extends JpaRepository <Visualization, Integer> {
+
+ 
 }
