+++ src/main/java/com/revature/app/service/VisualizationService.java
@@ -1,5 +1,14 @@
 package com.revature.app.service;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.revature.app.dao.VisualizationDao;
+
+@Service
 public class VisualizationService {
+	@Autowired
+	 VisualizationDao visualizationdao;
+	
 
 }
